+++ NEWS.md
@@ -3,10 +3,14 @@
 * Add new functions `calculate_rasmussen_2025_rf_cl_ccta_ptp`, 
 `calculate_rasmussen_2025_cacs_cl_ccta_ptp` to calculate pre-test probability of obstructive CAD from 2025 Rasmussen et. al. paper.
 
+* Add new functions `calculate_zuo_2025_predict_ocad_ptp`, 
+to calculate pre-test probability of obstructive CAD from 2025 Zuo et. al. paper.
+
 * Add new functions `calculate_miller_2023_vessel_50_cad_ptp`, 
 `calculate_miller_2023_lm_50_non_lm_70_cad_ptp` to calculate pre-test probability of obstructive CAD from 2023 Miller et. al. paper.
 
-* Add new functions `calculate_chen_2018_mfs_ptp` to calculate pre-test probability of obstructive CAD from 2018 Chen et. al. paper.
+* Add new functions `calculate_chen_2018_mfs_formula_ptp`, 
+`calculate_chen_2018_mfs_fig_3_ptp` to calculate pre-test probability of obstructive CAD from 2018 Chen et. al. paper.
 
 # pretestcad 1.1.0